Renamed error class and organized into error file.
