+++ src/types.ts
@@ -58,31 +58,3 @@ export type GetOwnedNftImageUrlFunction = (
   collectionAddress: string,
   tokenId: string
 ) => Promise<string>;
-
-export class VerificationError extends Error {
-  errorString?: string;
-
-  constructor(public statusCode: number, public label: string, error?: unknown) {
-    super(label);
-    this.name = "VerificationError";
-    if (error) {
-      this.errorString = error instanceof Error ? error.message : `${error}`;
-    }
-  }
-
-  get responseJson() {
-    return {
-      error: this.label,
-      ...(this.errorString && {
-        message: this.errorString,
-      }),
-    };
-  }
-}
-
-export class NotOwnerError extends Error {
-  constructor() {
-    super();
-    this.name = "NotOwnerError";
-  }
-}
